requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-swagger==1.0.16
aiosignal==1.2.0
async-lru==1.0.2
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
cryptography==36.0.1
dill==0.3.4
distro==1.6.0
frozenlist==1.3.0
genie==21.12
genie.libs.clean==21.12
genie.libs.conf==21.12
genie.libs.filetransferutils==21.12
genie.libs.health==21.12
genie.libs.ops==21.12
genie.libs.parser==21.12
genie.libs.sdk==21.12
gitdb==4.0.9
GitPython==3.1.18
idna==3.3
Jinja2==3.0.3
jsonpickle==2.1.0
junit-xml==1.9
MarkupSafe==2.0.1
multidict==6.0.2
netaddr==0.8.0
pathspec==0.9.0
prettytable==3.0.0
psutil==5.9.0
pyats==21.12
pyats-genie-command-parse==1.2.5
pyats.aereport==21.12
pyats.aetest==21.12
pyats.async==21.12
pyats.connections==21.12
pyats.datastructures==21.12
pyats.easypy==21.12
pyats.kleenex==21.12
pyats.log==21.12
pyats.reporter==21.12
pyats.results==21.12
pyats.tcl==21.12
pyats.topology==21.12
pyats.utils==21.12
pycparser==2.21
pyftpdlib==1.5.6
python-dotenv==0.19.2
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
smmap==5.0.0
tabulate==0.8.9
tftpy==0.8.0
tqdm==4.62.3
unicon==21.12
unicon.plugins==21.12
urllib3==1.26.8
wcwidth==0.2.5
xmltodict==0.12.0
yamllint==1.26.3
yarl==1.7.2
